+++ src/hlsync/__init__.py
@@ -1,3 +1,3 @@
 """Hook Line Sync."""
 
-__version__ = "0.8.30.9"
+__version__ = "0.8.30.11"